Reorganizing package: seperate timeseries fns from plotting ones
